backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlmodel==0.0.24
psycopg[binary]==3.2.6
pydantic-settings==2.9.1
